crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
feedparser==6.0.11
python-dotenv==1.0.1